LS1012ARDB: Change in script file for pfe_fw check
Add a check for pfe_fw in script used by flexbuild
for signing

+++ scripts/platforms/ls104x_1012_qspi.sh
@@ -46,7 +46,7 @@ if [ -f bootscript_enforce ]; then
     ./uni_sign input_files/uni_sign/ls104x_1012/input_initramfs_secure
 fi
 
-if [ -f pfe_fw ]; then
+if [ -f $FBDIR/$pfe_fw ]; then
     ./uni_sign input_files/uni_sign/ls104x_1012/input_pfe_secure
 fi
 
@@ -64,5 +64,7 @@ if [ -f secboot_hdrs_qspiboot.bin ]; then
     rm secboot_hdrs_qspiboot.bin
 fi
 touch secboot_hdrs_qspiboot.bin
+if [ -f $FBDIR/$pfe_fw ]; then
 dd if=hdr_pfe.out of=secboot_hdrs_qspiboot.bin bs=1K seek=1024
+fi
 dd if=hdr_kernel.out of=secboot_hdrs_qspiboot.bin bs=1K seek=2048
